+++ rspec-core/features/support/send_sigint_during_bisect.rb
@@ -12,8 +12,8 @@ module RSpec::Core::Formatters
       # Process.kill is not a synchronous call, so to ensure the output
       # below aborts at a deterministic place, we need to block here.
       # The sleep will be interrupted by the signal once the OS sends it.
-      # For the most part, this is only needed on JRuby, but we saw
-      # the asynchronous behavior on an MRI 2.0 travis build as well.
+      # For the most part, this is only needed on JRuby, but we have seen
+      # the asynchronous behavior on an MRI build as well.
       sleep 5
     end
   end