+++ rspec-core/lib/rspec/core/formatters/exception_presenter.rb
@@ -235,10 +235,6 @@ module RSpec
           ["Unable to find #{file_path} to read failed line"]
         rescue SnippetExtractor::NoSuchLineError
           ["Unable to find matching line in #{file_path}"]
-        rescue SecurityError
-          # :nocov: - SecurityError is no longer produced starting in ruby 2.7
-          ["Unable to read failed line"]
-          # :nocov:
         end
 
         def find_failed_line