+++ rspec-core/spec/rspec/core/memoized_helpers_spec.rb
@@ -647,11 +647,4 @@ module RSpec::Core
       end
     end
   end
-
-  RSpec.describe 'Module#define_method' do
-    it 'retains its normal private visibility on Ruby versions where it is normally private', :skip => RUBY_VERSION >= '2.5' do
-      a_module = Module.new
-      expect { a_module.define_method(:name) { "implementation" } }.to raise_error NoMethodError
-    end
-  end
 end