+++ rspec-expectations/spec/rspec/matchers/built_in/exist_spec.rb
@@ -134,11 +134,9 @@ RSpec.describe "exist matcher" do
         end
       end
 
-      if RUBY_VERSION > "1.9"
-        context 'Dir has deprecated exists?' do
-          it 'will not call exists? triggering the warning' do
-            expect(Dir).to exist Dir.pwd
-          end
+      context 'Dir has deprecated exists?' do
+        it 'will not call exists? triggering the warning' do
+          expect(Dir).to exist Dir.pwd
         end
       end
     end