+++ rspec-support/spec/rspec/support/source_spec.rb
@@ -87,18 +87,11 @@ module RSpec::Support
       it 'returns a hash containing nodes for each line number' do
         expect(source.nodes_by_line_number).to match(
           1 =>
-            if RUBY_VERSION >= '2.6.0'
-              [
-                an_object_having_attributes(:type => :@int),
-                an_object_having_attributes(:type => :@period),
-                an_object_having_attributes(:type => :@ident)
-              ]
-            else
-              [
-                an_object_having_attributes(:type => :@int),
-                an_object_having_attributes(:type => :@ident)
-              ]
-            end,
+            [
+              an_object_having_attributes(:type => :@int),
+              an_object_having_attributes(:type => :@period),
+              an_object_having_attributes(:type => :@ident)
+            ],
           2 => [
             an_object_having_attributes(:type => :@ident),
             an_object_having_attributes(:type => :@ident)